backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.1
